+++ .eslintrc.json
@@ -55,7 +55,7 @@
     "arrow-spacing": 2,
     "brace-style": 2,
     "no-useless-constructor": 2,
-    "no-multi-spaces": 0,
+    "no-multi-spaces": 2,
     "no-nested-ternary": 2,
     "prefer-const": 2,
     "react/jsx-tag-spacing": [
