+++ App.js
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { StatusBar }  from 'react-native';
+import { StatusBar } from 'react-native';
 import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
 import SplashScreen from 'react-native-splash-screen';
