+++ babel.config.js
@@ -5,23 +5,16 @@ module.exports = {
       'module-resolver',
       {
         alias: {
-          screens: './src/screens',
-          components: './src/components',
-          // assets: './src/assets',
-          // httpClient: './src/httpClient',
-          // services: './src/services',
-          // components: './src/components',
-          // constants: './src/constants',
-          // screens: './src/screens',
-          // hooks: './src/hooks',
-          // locale: './src/locale',
-          // reducers: './src/reducers',
-          // selectors: './src/selectors',
-          // store: './src/store',
-          // utils: './src/utils',
-          // navigators: './src/navigators',
-          // storybook: './storybook',
-          // validations: './src/validations',
+          '@screens': './src/screens',
+          '@components': './src/components',
+          '@actions': './src/reduxConfig/actions',
+          '@reduxConstants': './src/reduxConfig/actions/constants',
+          '@reducers': './src/reduxConfig/reducers',
+          '@sagas': './src/reduxConfig/sagas',
+          '@hooks': './src/hooks',
+          '@services': './src/services',
+          '@helpers': './src/helpers',
+          '@navigators': './src/navigation',
         },
       },
     ],
