--- src/reduxConfig/appReducer/constants.js
@@ -1,4 +1,0 @@
-/* eslint-disable no-multi-spaces */
-
-export const FETCH_DATA = 'APP_REDUCER::FETCH_DATA';
-export const SET_DATA   = 'APP_REDUCER::SET_DATA';
