--- src/reduxConfig/authReducer/constants.js
@@ -1,7 +1,0 @@
-// AUTH REDUCER
-export const SIGN_UP          = 'AUTH_REDUCER::SIGN_UP';
-export const SIGN_IN          = 'AUTH_REDUCER::SIGN_IN';
-export const SIGN_OUT         = 'AUTH_REDUCER::SIGN_OUT';
-export const SIGN_UP_REQUEST  = 'AUTH_REDUCER::SIGN_UP_REQUEST';
-export const SIGN_IN_REQUEST  = 'AUTH_REDUCER::SIGN_IN_REQUEST';
-export const SIGN_OUT_REQUEST = 'AUTH_REDUCER::SIGN_OUT_REQUEST';
