+++ src/reduxConfig/sagas/index.js
@@ -1,5 +1,5 @@
-import appSagas from '../appReducer/sagas';
-import authSagas from '../authReducer/sagas';
+import appSagas from '@sagas/appSagas';
+import authSagas from '@sagas/authSagas';
 
 export default [
   ...appSagas,
