+++ src/screens/Login/index.js
@@ -6,13 +6,13 @@ import {
 } from 'react-native';
 import { useDispatch } from 'react-redux';
 
-import Button from 'components/Button';
-import Input from 'components/Input';
+import Button from '@components/Button';
+import Input from '@components/Input';
 import useSetNavigationOptions from '../../hooks/useSetNavigationOptions';
 
 import { useNavigation } from '@react-navigation/native';
 
-import * as authActions from '../../reduxConfig/authReducer/actions';
+import authActions from '@actions/authActions';
 
 import styles from './styles';
