+++ src/screens/Welcome/index.js
@@ -11,7 +11,7 @@ import useSetNavigationOptions from '../../hooks/useSetNavigationOptions';
 
 import { useNavigation } from '@react-navigation/native';
 
-import * as authActions from '../../reduxConfig/authReducer/actions';
+import authActions from '@actions/authActions';
 
 import styles from './styles';
